refactor(ui): api client + timeline module (#343 phase A)
- client.ts: BASE_URL, BASE_API_URL, axios timeout
- timeline.ts: timeline/unknowns fetch+export + UnknownDetection type
- api.tsx: re-exports; drop duplicate timeline/unknowns block
- queryKeys: timeline, unknowns, corrections, overview, calendar, species.directory
- Timeline/Unknowns: use centralized query keys
- Vitest: timeline.test.ts

Made-with: Cursor
