+++ app/ui/src/api/client.ts
@@ -1,0 +1,11 @@
+/** HTTP client defaults and base URLs for UI → Hub API (#343). */
+import axios from 'axios';
+
+// Relative path = same origin (works with any host/IP). При SSR/тестах — из env или дефолт.
+export const BASE_URL =
+  typeof window !== 'undefined'
+    ? ''
+    : (import.meta.env?.VITE_BASE_URL as string) || '';
+export const BASE_API_URL = `${BASE_URL}/api/ui`;
+
+axios.defaults.timeout = 30000;
