+++ app/ui/src/api/queryKeys.ts
@@ -1,4 +1,4 @@
-/** Central React Query keys (#296) — invalidate via queryKeys.* for stable contracts. */
+/** Central React Query keys (#296, #343) — invalidate via queryKeys.* for stable contracts. */
 
 export const queryKeys = {
   settings: {
@@ -9,6 +9,16 @@ export const queryKeys = {
   },
   species: {
     observed: ['species', 'observed'] as const,
+    /** Справочник видов (Unknowns picker и др.). */
+    directory: ['species'] as const,
+  },
+  overview: {
+    all: ['overview'] as const,
+  },
+  calendar: {
+    /** Кэш вкладок/виджетов, привязанных к строке «timeline». */
+    timelineTab: ['timeline'] as const,
+    migration: ['migration-calendar'] as const,
   },
   system: {
     readiness: ['system', 'readiness'] as const,
@@ -18,5 +28,30 @@ export const queryKeys = {
     visitors: (days: number) => ['system', 'visitors', days] as const,
     processorLogs: (lines: number) =>
       ['system', 'processorLogs', lines] as const,
+  },
+  /** Таймлайн + счётчик unknowns на той же дате (#343). */
+  timeline: {
+    observerTimezone: ['timeline-observer-timezone'] as const,
+    speciesVisits: (
+      date: string,
+      timeOfDay: string,
+      filterHour: number | null,
+    ) => ['speciesVisits', date, timeOfDay, filterHour] as const,
+    /** Префикс для invalidateQueries — все окна таймлайна. */
+    speciesVisitsAll: ['speciesVisits'] as const,
+    unknownsCount: (
+      date: string,
+      timeOfDay: string,
+      filterHour: number | null,
+    ) => ['unknowns-count', date, timeOfDay, filterHour] as const,
+    unknownsCountAll: ['unknowns-count'] as const,
+  },
+  unknowns: {
+    list: (date: string, timeOfDay: string) =>
+      ['unknowns', date, timeOfDay] as const,
+    all: ['unknowns'] as const,
+  },
+  corrections: {
+    recent: ['corrections-recent'] as const,
   },
 } as const;
